Fix: Remove unnecessary .placeholder file
Removed the `.placeholder` file from the `js/ui` directory as it is no longer needed after adding actual module files to the directory.
